requirements.txt
beautifulsoup4==4.10.0
llama-cpp-python==0.1.74
langchain==0.0.257
python-telegram-bot==20.4
PyYAML==6.0
typer==0.9.0
arxiv
pymupdf
tiktoken
